backend services + removing BigDecimal types + introducing dates types

+++ backend/app/src/main/java/eu/viandeendirect/repository/OrderItemRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface OrderItemRepository extends CrudRepository<OrderItem, Long> {
+public interface OrderItemRepository extends CrudRepository<OrderItem, Integer> {
     @Query("SELECT i FROM OrderItem i WHERE i.order = :order")
     List<OrderItem> findByOrder(@Param("order") Order order);
 }
