+++ backend/app/src/test/java/eu/viandeendirect/service/ProducerServiceForTests.java
@@ -5,8 +5,6 @@ import eu.viandeendirect.service.specs.ProducerServiceSpecs;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
 
-import java.math.BigDecimal;
-
 @Service
 @Profile("test")
 public class ProducerServiceForTests implements ProducerServiceSpecs {
@@ -14,7 +12,7 @@ public class ProducerServiceForTests implements ProducerServiceSpecs {
     @Override
     public Producer getAuthenticatedProducer() {
         var producer = new Producer();
-        producer.setId(BigDecimal.valueOf(1000));
+        producer.setId(1000);
         return producer;
     }
 }
